helper: debug: move odph_debug.h into helper API
Moved the debug header to be part of the helper API.
