+++ helper/cuckootable.c
@@ -44,8 +44,8 @@
 #include <errno.h>
 #include <stdio.h>
 
-#include "odp/helper/odph_cuckootable.h"
-#include "odph_debug.h"
+#include <odp/helper/odph_cuckootable.h>
+#include <odp/helper/odph_debug.h>
 #include <odp_api.h>
 
 /* More efficient access to a map of single ullong */
