+++ helper/iplookuptable.c
@@ -12,8 +12,8 @@
 #include <stdio.h>
 
 #include <odp/helper/odph_iplookuptable.h>
+#include <odp/helper/odph_debug.h>
 #include "odph_list_internal.h"
-#include "odph_debug.h"
 #include <odp_api.h>
 
 /** @magic word, write to the first byte of the memory block
