+++ helper/linux/thread.c
@@ -24,7 +24,7 @@
 #include <odp_api.h>
 #include <odp/helper/linux/pthread.h>
 #include <odp/helper/linux/process.h>
-#include "odph_debug.h"
+#include <odp/helper/odph_debug.h>
 
 static void *_odph_run_start_routine(void *arg)
 {
