+++ helper/test/cuckootable.c
@@ -50,7 +50,6 @@
 #include <time.h>
 
 #include <odp_api.h>
-#include <odph_debug.h>
 #include <odp/helper/odph_api.h>
 
 /*******************************************************************************
