+++ helper/test/iplookuptable.c
@@ -13,9 +13,7 @@
 #include <errno.h>
 
 #include <odp_api.h>
-#include <odph_debug.h>
 #include <odp/helper/odph_api.h>
-#include <odp/helper/ip.h>
 
 static void print_prefix_info(
 		const char *msg, uint32_t ip, uint8_t cidr)
