+++ helper/test/parse.c
@@ -6,8 +6,6 @@
 
 #include "config.h"
 
-#include <odph_debug.h>
-
 #include <odp_api.h>
 #include <odp/helper/odph_api.h>
